requirements.txt
argh==0.26.2
backoff==1.3.1
blessings==1.6
bpython==0.15
curtsies==0.2.9
decorator==4.0.10
greenlet==0.4.10
ipify==1.0.0
ipython==5.1.0
ipython-genutils==0.1.0
pathtools==0.1.2
peewee==2.8.3
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.8
ptyprocess==0.5.1
Pygments==2.1.3
pyinotify==0.9.6
PyQt5==5.7
python-dateutil==2.5.3
PyYAML==3.12
requests==2.11.1
simplegeneric==0.8.1
sip==4.18.1
six==1.10.0
traitlets==4.3.1
watchdog==0.8.3
wcwidth==0.1.7
